+++ grain/_src/python/dataset/transformations/prefetch_test.py
@@ -726,6 +726,92 @@ class MultiprocessPrefetchIterDatasetTest(parameterized.TestCase):
     with self.assertRaises(Exception):
       list(ds)
 
+  def test_multiprocess_prefetch_with_sequential_slice(self):
+    ds = dataset.MapDataset.source(range(10)).to_iter_dataset()
+    ds = prefetch.MultiprocessPrefetchIterDataset(
+        ds,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    self.assertEqual(list(ds), [0, 4, 7, 1, 5, 8, 2, 6, 9, 3])
+
+  def test_multiprocess_prefetch_with_default_slice_non_sequential(self):
+    ds = dataset.MapDataset.source(range(10)).to_iter_dataset()
+    ds_sequential_off = prefetch.MultiprocessPrefetchIterDataset(
+        ds,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=False,
+    )
+    ds_sequential_default = prefetch.MultiprocessPrefetchIterDataset(
+        ds,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+    )
+    elements_sequential_off = list(ds_sequential_off)
+    elements_sequential_default = list(ds_sequential_default)
+    self.assertEqual(
+        elements_sequential_off,
+        elements_sequential_default,
+    )
+    self.assertEqual(
+        [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+        elements_sequential_default,
+    )
+
+  def test_multiprocess_prefetch_sequential_slice_order_from_source(self):
+    ds = dataset.MapDataset.source(range(10)).to_iter_dataset()
+    ds_sequential_on = prefetch.MultiprocessPrefetchIterDataset(
+        ds,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    elements_sequential_on = list(ds_sequential_on)
+    self.assertEqual([0, 4, 7, 1, 5, 8, 2, 6, 9, 3], elements_sequential_on)
+
+  def test_multiprocess_prefetch_sequential_slice_order_from_range(self):
+    ds_range = dataset.MapDataset.range(10).to_iter_dataset()
+    ds_range_sequential_on = prefetch.MultiprocessPrefetchIterDataset(
+        ds_range,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    elements_range_sequential_on = list(ds_range_sequential_on)
+    self.assertEqual(
+        [0, 4, 7, 1, 5, 8, 2, 6, 9, 3],
+        elements_range_sequential_on,
+    )
+
+  def test_multiprocess_prefetch_sequential_slice_order_from_range_slice(self):
+    ds_range = dataset.MapDataset.range(
+        start=2, stop=21, step=3
+    ).to_iter_dataset()
+    ds_range_sequential_on = prefetch.MultiprocessPrefetchIterDataset(
+        ds_range,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    elements_range_sequential_on = list(ds_range_sequential_on)
+    self.assertEqual(
+        [2, 11, 17, 5, 14, 20, 8],
+        elements_range_sequential_on,
+    )
+
+  def test_multiprocess_prefetch_sequential_slice_order_same(self):
+    ds_source = dataset.MapDataset.source(range(10)).to_iter_dataset()
+    ds_range = dataset.MapDataset.range(10).to_iter_dataset()
+    ds_source_mp = prefetch.MultiprocessPrefetchIterDataset(
+        ds_source,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    ds_range_mp = prefetch.MultiprocessPrefetchIterDataset(
+        ds_range,
+        options.MultiprocessingOptions(num_workers=3, per_worker_buffer_size=1),
+        sequential_slice=True,
+    )
+    elements_source = list(ds_source_mp)
+    elements_range = list(ds_range_mp)
+    self.assertEqual(elements_source, elements_range)
+
   def test_options_after_prefetch(self):
     ds = dataset.MapDataset.source([1, 2, 3]).repeat(1000)
     ds = ds.filter(lambda x: x > 2)